Moved snaphooks into snapstate/hooks.
